move generate graph function out of student report add smoke test for it
