+++ spec/javascripts/student_profile/GenerateGraph.test.js
@@ -1,0 +1,7 @@
+import generateGraph from '../../../app/assets/javascripts/student_profile/pdf/GenerateGraph';
+
+describe('generateGraph', () => {
+  it('generates graph without crashing', () => {
+    const result = generateGraph("#graph", "Attendence Details", {}, "Graph", []);
+  });
+});
